requirements.txt
asn1crypto==0.24.0
astroid==2.2.5
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.3
djangorestframework==3.10.1
docker==4.0.2
idna==2.8
isort==4.3.20
lazy-object-proxy==1.4.1
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pycparser==2.19
pylint==2.3.1
pyOpenSSL==19.0.0
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
urllib3==1.25.3
websocket-client==0.56.0
wrapt==1.11.1
